chore(graphql): refactor
