+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/graphql/users/resolvers.py
@@ -1,5 +1,5 @@
 from django.contrib.auth import get_user_model
 
 
-def resolve_users(info):
+def get_all_users(info):
     return get_user_model().objects.all()
